Create README.md
What can be better than a cup of coffee during a break? A coffee that you don’t have to make yourself. It’s enough to press a couple of buttons on the machine and you get a cup of energy; but first, we should teach the machine how to do it. In this project, I worked on programming a coffee machine simulator. The machine works with typical products: coffee, milk, sugar, and plastic cups; if it runs out of something, it shows a notification. You can get three types of coffee: espresso, cappuccino, and latte. Since nothing’s for free, it also collects money.

This project I made while learning Java at JetBrains Academy. It demonstrates knowledge of the basic OOP, and its main concepts such as classes, class methods, and attributes. It also illustrates work with methods, loops, and conditions.